drop mainchain vs sidechain distinction from OMOL format, and improve Osprey/Molscope interop
The molecule format itself is too general to require sub-residue atom partitioning, and the heuristics to do it are too brittle to automatically apply everywhere. We can do mainchain/sidechain partitioning later when we actually need it (eg, in the GUI for rendering, or in the eventual conf space format when we need to determine how to apply mutations).