WIP: refactor design positions to allow an assignment system
That is, allow setting conformations at design positions on copies of molecules.
Useful so we can parallelize the conf space compiler and speed it up.

The current compiler code is still too complicated, with all the molecule/atom mapping madness, so more work needs to be done to simplify that.

Also, the GUI itself hasn't been updated to modify molecule copies yet.