use concurrency to speed up the conf space compiler
The compiler spends most of its time waiting for blocking operations, like OspreyService requests, so there's quite a bit of speedup to be had just by making those calls asynchronous. This lets the compiler prep the next few calls while it waits for the first calls to finish, like a pipeline.

This required a few changes:
 * expose the natural concurrency in OspreyService's KTor client, powered by the amazing Kotlin coroutines
 * make callers of those functions `suspend`
 * at the top of the call chain (eg in GUIs, the compiler), use `runBlocking` to launch the coroutines and wait for them to finish
 * make the higher-order functions of `suspend` functions `inline`, so we don't have to make `suspend` and non-`suspend` copies of those functions
 * add a simple liming mechanism to keep the coroutine engine from trying to run all the coroutines at the same time and running the JVM out of memory